First try at NiceGUI
This was more getting the UI how i wanted it to look like. Got the log viewer working as a bonus haha. I am still not sure if i will end up going with NiceGUI or something else. But as of not, i am liking the simplicity of things. Not super simple but also not creating things from scratch with HTML,CSS,JS complicated either. idk, still a WIP and will see what the future holds :)

Dockerfile.dev is a temp file and will be deleted in prod.