Fix season_id being null/None
Fixed an issue where the season ID could be null/None and crash the MainLoop thread. MDNX listed the header as "Season: 4" but episode lines as "- Season 43 -," for Link Click (GP5HJ8E81). So, the parser in process_console_output() created a new placeholder season (example, S5) with None as the season_id and season_name (in JSON, these turn into null). At this stage, we has S4 with the correct season_id and season_name, but no episodes. Then we had S5 with null as the season_id/name but with correct episodes. Then, the prune/renumber step dropped/shifted S4 because it didn't have any episodes, leaving the placeholder with null values as the active season with a null season_id.