Remove REMOVE_ALL_ACTIVE_STREAMS option
The reason why this was added was because even though this option was suppost to be able to be set in cli-defaults.yml, it never worked. After testing, it seems like multi-downloader-nx is honoring the defaults file, so this additional feature is no longer needed. If you were using this, set the "tsd" value to "true".