Add defaults
If you dont have a certain key-value pair, instead of erroring out when trying to access the variable, it will use the defaults. Made it so it merges the default settings into the settings giving by the user. Any settings coming from the user is used to overwrite the default settings (closes #50)