Trigger media server scan only when changes exist
Media server scan is now triggered only if there are new items in the notifications buffer and the media server is configured. This prevents unnecessary scans when there are no new changes. We can rely on the notifications buffer since any time something is downloaded or changed because of a new dub/sub, it is added to the buffer. If nothing is in the buffer, nothing has changed, hence no library scan necessary.