Plex auth token and config.json saving fixes
I though that when in the webui, go to a episodes details and click on "View XML", the plex token there was the token you can use for stuff. Well, i came to the realization that that token is a temporary token. So, the changes in the MediaServerManager.py module aims to use a user auth token. This requires the user to open a URL in their web web browser and auth manually. I have set the timeout to be 10 minutes for now.

The changes in update_app_config() is for when saving new keys or changing values of a key, it would ruin the formatting/order of the JSON file. I have fixed this by getting the order from CONFIG_DEFAULTS and using that as a reference.