Add fuzzy finding
When determining if an episode is download, it only did a simple "hey, does this path exist?" (os.path.exists).
While it still does that, it can also fuzzy find the series and mark an episode as already downloaded. This comes in handy is your folder is marked as "chars adventures" but mdnx-auto-dl is looking for "char's adventures".

This is enabled by default and can be disabled by "FUZZY_MATCHING_ENABLED" to "false".