Fix fuzzy matching
Would match series and seasons names correctly, but would fail to download new episodes because it would be matching the wrong episode (example is it saying it fuzzy matched S01E01 to not yet downloaded S01E04, so it marks it as already downloaded).

Also made the default threshold 90 instead of 92.